// packages/core/src/message.d.ts
export type MessageHeaders = Headers;
/**
 * {@link Metadata} type for chat messages.
 */
export type MessageMetadata = Metadata;
/**
 * {@link OperationMetadata} type for a chat messages {@link Operation}.
 */
export type MessageOperationMetadata = OperationMetadata;
/**
 * Represents the detail of a message deletion or update.
 */
export interface Operation {
    /**
     * The optional clientId of the user who performed the update or deletion.
     */
    clientId?: string;
    /**
     * The optional description for the update or deletion.
     */
    description?: string;
    /**
     * The optional metadata associated with the update or deletion.
     */
    metadata?: MessageOperationMetadata;
}
// ...
export declare class DefaultMessage implements Message {
    readonly serial: string;
    readonly clientId: string;
    readonly roomId: string;
    readonly text: string;
    readonly metadata: MessageMetadata;
    readonly headers: MessageHeaders;
    readonly action: ChatMessageActions;
    readonly version: string;
    readonly createdAt: Date;
    readonly timestamp: Date;
    readonly operation?: Operation | undefined;
    constructor(serial: string, clientId: string, roomId: string, text: string, metadata: MessageMetadata, headers: MessageHeaders, action: ChatMessageActions, version: string, createdAt: Date, timestamp: Date, operation?: Operation | undefined);
    get isUpdated(): boolean;
    get isDeleted(): boolean;
    get updatedBy(): string | undefined;
    get deletedBy(): string | undefined;
    get updatedAt(): Date | undefined;
    get deletedAt(): Date | undefined;
    versionBefore(message: Message): boolean;
    versionAfter(message: Message): boolean;
    versionEqual(message: Message): boolean;
    before(message: Message): boolean;
    after(message: Message): boolean;
    equal(message: Message): boolean;
}